Bug fix: multiple pipeline collision
When multiple pipelines share the same rev #'s, the lookup for diffs would try across both pipelines because that was the only hash.  Now we use the pipeline's name and revision (label) number.   The hash is because we are looking at **per pipeline** diffs.  The name gives a unique hash for each pipeline and the revision (label) number gives a unique hash among the same pipeline but different times.